youtube_monitor/requirements.txt
# YouTube Playlist Monitor Requirements

# Core dependencies
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
youtube-transcript-api>=0.6.1

# HTTP requests
requests>=2.31.0

# Environment management
python-dotenv>=1.0.0

# Development and testing
pytest>=8.2.0
pytest-asyncio>=0.23.0

# Utility libraries
click>=8.1.7
rich>=13.8.0 